Remove image URL manipulation
removeProtocol should not be required. This was introduced in 2017 when both
HTTP & HTTPS were in general use. Nowadays, only HTTPS is in general use for
ecommerce websites. Browsers will typically display passive mixed content such
as images. If websites are having trouble with mixed content this will not be
only because of Algolia search.
https://developer.mozilla.org/en-US/docs/Web/Security/Mixed_content/How_to_fix_website_with_mixed_content

removeDoubleSlashes is introducing a bug: a URL which is a full path only is
having an additional slash added to the start, producing a valid-but-wrong URL.
For example, '/media//wysiwyg//search.png' becomes '//media/wysiwyg/search.png'.
Browsers cope just fine with URLs containing multiple slashes, so no
manipulation is required here.