Fix potential RequireJS race condition
In some cases, RequireJS may execute the `recommend.js` callback before `window.objectId` is set in `algolia/algoliasearch-magento-2/view/frontend/templates/recommend/products.phtml`. When this happens, an *Uncaught ReferenceError* is thrown and short circuits RequireJS from loading the rest of the dependent scripts. In our case, that meant that the callback in `autocomplete.js` never loaded, which caused the search box to not render. The error condition is also more likely when browser caching is enabled, since the `recommend.js` script is then more likely to be loaded from the cache and execute sooner.

This fix wraps the `recommend.js` callback in a jQuery ready callback so it does not execute until after pageload.

Console error for reference:
```
recommend.js:16 Uncaught ReferenceError: objectId is not defined
    at recommend.js:16:58
    at Object.execCb (require.js:1696:33)
    at Module.check (require.js:883:51)
    at Module.<anonymous> (require.js:1139:34)
    at require.js:134:23
    at require.js:1189:21
    at each (require.js:59:31)
    at Module.emit (require.js:1188:17)
    at Module.check (require.js:938:30)
    at Module.enable (require.js:1176:22)
```