Update Configuration.php
typecast the $query variable to comply with the PHP 8 standards of [html_entity_decode](https://www.php.net/manual/en/function.html-entity-decode.php)

If the search results page (`/catalogsearch/result`) is requested and the `q` query parameter is not present the page will error.  The code errors when generating the `algoliaConfig` JS variable which means that Algolia can not be instantiated.  This means no products show on the page - the expected result is that a blank search query would be performed and thus a set of products would be displayed.